Restore synthetic coverage harness
