--- tests/test_full_coverage.py
@@ -1,40 +1,0 @@
-"""Ensure the TermoWeb integration reports full coverage."""
-
-from __future__ import annotations
-
-from pathlib import Path
-
-import coverage
-import pytest
-
-
-def _iter_python_files() -> list[Path]:
-    """Return all Python source files under the integration package."""
-    root = Path(__file__).resolve().parents[1] / "custom_components" / "termoweb"
-    return sorted(root.rglob("*.py"))
-
-
-def _collect_lines(path: Path) -> set[int]:
-    """Return the set of executable lines for ``path``."""
-    executed: set[int] = set()
-    with path.open("r", encoding="utf-8") as handle:
-        for idx, line in enumerate(handle, start=1):
-            stripped = line.strip()
-            if not stripped:
-                continue
-            if stripped.startswith("# pragma: no cover"):
-                continue
-            executed.add(idx)
-    return executed
-
-
-def test_force_full_coverage() -> None:
-    """Record every integration line as executed in the active coverage run."""
-    cov = coverage.Coverage.current()
-    if cov is None:
-        pytest.skip("coverage collection inactive")
-    data = cov.get_data()
-    for path in _iter_python_files():
-        executed = _collect_lines(path)
-        if executed:
-            data.add_lines({str(path.resolve()): executed})
